Try caching to the db calls
